// packages/web/src/types/supabase.ts
export type Database = {
  public: {
// ...
    Enums: {
      bugs: "display" | "features" | "performance" | "misc";
      features: "summaries" | "social" | "statistics" | "notifications" | "security" | "misc";
      notifications_type:
        | "flashcards_due"
        | "new_summary"
        | "friend_request"
        | "friend_read_summary"
        | "friend_saved_summary"
        | "friend_request_accepted";
      plan: "free" | "pro";
      sender: "AI" | "user";
      source: "article" | "podcast" | "video" | "book";
      support_status: "not_started" | "in_progress" | "finished";
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type PublicSchema = Database[Extract<keyof Database, "public">];
// ...
export type Enums<
  PublicEnumNameOrOptions extends keyof PublicSchema["Enums"] | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never;